requirements.txt
Flask==2.1.3
psycopg2-binary==2.9.3
Scrapy==2.6.1
selenium==4.3.0